wpi(4): use more correct types.
This change fixes some amount of -Wsign-conversion and -Wconversion warnings
and sets correct sizes for some variables (as a result, some loop counters
were touched too).

Submitted by:	<s3erios@gmail.com>
Differential Revision:	https://reviews.freebsd.org/D3763

+++ sys/dev/wpi/if_wpireg.h
@@ -978,7 +978,7 @@ static const struct wpi_pmgt {
 	uint32_t	rxtimeout;
 	uint32_t	txtimeout;
 	uint32_t	intval[5];
-	int		skip_dtim;
+	uint8_t		skip_dtim;
 } wpi_pmgt[WPI_NDTIMRANGES][WPI_NPOWERLEVELS] = {
 	/* DTIM <= 10 */
 	{
